# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "XMeme API",
        "description": "API for XMEME appplication",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/memes": {
            "get": {
                "summary": "Get list of 100 memes in reverse chronological order",
                "description": "Returns list of 100 recent memes posted",
                "responses": {
                    "200": {
                        "description": "List of memes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    },
                                    "caption": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error in fetching memes"
                    }
                }
            },
            "post": {
                "summary": "API to post a meme",
                "description": "Posts a meme",
                "responses": {
                    "201": {
                        "description": "Successfully posted the meme",
                        "schema": {
                            "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                }
                        }
                    },
                    "409": {
                        "description": "Posted Duplicate meme"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Server Error in fetching memes"
                    }
                }
            }
        },
        "/memes/{id}": {
            "get": {
                "summary": "Get the meme post identified by the unique id",
                "description": "Get the meme post identified by the unique id",
                "responses": {
                    "200": {
                        "description": "Post",
                        "schema": {
                            "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    },
                                    "caption": {
                                        "type": "string"
                                    }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error in fetching meme post"
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "The meme post with the particular id couldn't be found"
                    }
                }
            },
            "patch": {
                "summary": "Update the url and caption of meme",
                "description": "Update the url and caption of meme",
                "responses": {
                    "200": {
                        "description": "Posts a meme",
                        "schema": {
                            "type": "object",
                                "properties": {
                                    "url" : {
                                        "type": "string"
                                    },
                                    "caption": {
                                        "type": "string"
                                    }
                                }
                        }
                    },
                    "409": {
                        "description": "Posted Duplicate meme"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Server Error in fetching memes"
                    }
                }
            }
        },
        "/like/{id}": {
            "get": {
                "summary": "Get the likes for meme post",
                "description": "Get the likes for meme post",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                                "properties": {
                                "likes": "integer"
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error in fetching likes"
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "The meme post with the particular id couldn't be found"
                    }
                }
            },
            "post": {
                "summary": "Add a like to meme post",
                "description": "Add a like to meme post",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Id not found"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Server Error in fetching memes"
                    }
                }
            }
        },
        "/unlike/{id}": {
            "post": {
                "summary": "Remove like from meme post",
                "description": "Remove like from meme post",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Id not found"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Server Error in fetching memes"
                    }
                }
            }
        }

    },
    "tags": [
        {
            "name": "Memes",
            "description": "API for Meme posts made"
        },
        {
            "name": "Like",
            "description": "Api for Likes made"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ]
}